Add profiler to Experiment
